src/server/requirements.txt
boto3==1.13.10
Django==3.2.25
django-cors-headers==3.2.1
django-storages==1.9.1
psycopg2==2.8.5
python-magic==0.4.27
djangorestframework==3.12.4
djangorestframework-gis==0.17
django-leaflet==0.28.1
django-admin-rangefilter==0.8.6
django-redis==5.0.0
geopy==2.2.0
gunicorn==23.0.0
pandas==2.2.0
scikit-learn==1.5.0
shapely==2.0.2
strawberry-graphql==0.209.7
strawberry-graphql-django==0.21.0
wagtail==5.2.6
wagtail-footnotes==0.10.0
wagtail-headless-preview==0.7.0
